# lxdweb/docs/swagger.yaml
basePath: /
...
host: localhost:3000
info:
  contact:
    email: xkatld@example.com
    name: xkatld
    url: https://github.com/xkatld/zjmf-lxd-server
  description: 基于 Session 的认证
  license:
    name: MIT
    url: https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE
  termsOfService: https://github.com/xkatld/zjmf-lxd-server
  title: LXD Web 管理平台
  version: 1.0.3
paths:
...
  /api/containers:
    get:
      description: 从本地数据库获取所有容器缓存信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回容器列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 获取容器列表
      tags:
      - 容器管理
  /api/containers/{name}:
    get:
      description: 获取指定容器的详细信息
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回容器详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点或容器不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取容器详细信息
      tags:
      - 容器管理
  /api/containers/{name}/delete:
    post:
      description: 删除指定的容器并清理数据库缓存
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 删除容器
      tags:
      - 容器管理
  /api/containers/{name}/password:
    post:
      consumes:
      - application/json
      description: 重置指定容器的root密码
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 密码参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 重置容器密码
      tags:
      - 容器管理
  /api/containers/{name}/refresh:
    post:
      description: 实时获取指定容器的最新信息并更新缓存
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 刷新单个容器信息
      tags:
      - 容器管理
  /api/containers/{name}/reinstall:
    post:
      consumes:
      - application/json
      description: 重装指定容器的操作系统
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 重装参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 重装成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 重装容器系统
      tags:
      - 容器管理
  /api/containers/{name}/restart:
    post:
      description: 重启指定的容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 重启成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 重启容器
      tags:
      - 容器管理
  /api/containers/{name}/start:
    post:
      description: 启动指定的容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 启动成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 启动容器
      tags:
      - 容器管理
  /api/containers/{name}/stop:
    post:
      description: 停止指定的容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 停止成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 停止容器
      tags:
      - 容器管理
  /api/containers/{name}/suspend:
    post:
      description: 暂停指定的容器（Frozen状态）
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 暂停成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 暂停容器
      tags:
      - 容器管理
  /api/containers/{name}/traffic/reset:
    post:
      description: 重置指定容器的流量统计
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 重置容器流量
      tags:
      - 容器管理
  /api/containers/{name}/unsuspend:
    post:
      description: 恢复被暂停的容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 恢复成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 恢复容器
      tags:
      - 容器管理
  /api/containers/cache:
    get:
      description: 从本地数据库缓存获取容器列表，支持按node_id筛选
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回容器列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 从本地数据库缓存获取容器列表
      tags:
      - 容器管理
  /api/containers/create:
    post:
      consumes:
      - application/json
      description: 在指定节点创建新的LXD容器
      parameters:
      - description: 容器配置参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 创建容器
      tags:
      - 容器管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"